Video: OV camera: Enable camera config for power investigation
Note: For performance reason, we should build-in those driver as
much as possible, but so far, we'd like to set them as module if
possible, the reason is that we can easily debug power issue for
finding culprit with those load/unload way so that we don't need
to compile and re-burn the kernel again and again.
Before product level release, we should revert such changes and
make them 'Y' as much as possible.

Signed-off-by: Austin Zhang <austin.zhang@intel.com>