atomisp2: set initial streaming state to disabled
If user space application fails to explicitly disable streaming before
closing the device, the streaming state will be left with an incorrect
value. This will possible make it impossible to reopen the device.
This fixes an issue where after camera application has crashed, the
camera device can not be opened any more.

Signed-off-by: Yan Feilong <feilongx.yan@intel.com>